+++ backend/src/modules/appointments/infra/http/controllers/AppointmentsController.ts
@@ -1,0 +1,24 @@
+import { Request, Response } from 'express'
+import { parseISO } from 'date-fns'
+import { container as dependencyInjector } from 'tsyringe'
+
+import CreateAppointmentService from '@/modules/appointments/services/CreateAppointmentService'
+
+export default class AppointmentsController {
+
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { provider_id, date } = request.body
+
+    const parsedDate =  parseISO(date)
+
+    const createAppointmentService = dependencyInjector.resolve(CreateAppointmentService)
+
+    const appointment = await createAppointmentService.execute({
+      provider_id,
+      date: parsedDate
+    })
+
+    return response.json(appointment)
+  }
+
+}
