+++ backend/src/modules/users/infra/http/controllers/SessionsController.ts
@@ -1,0 +1,18 @@
+import { Request, Response } from 'express'
+import { container as dependencyInjector } from 'tsyringe'
+
+import AuthenticateUserService from '@/modules/users/services/AuthenticateUserService'
+
+export default class SessionsController {
+
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { email, password } = request.body
+
+    const authenticateUserService = dependencyInjector.resolve(AuthenticateUserService)
+
+    const { user, token } = await authenticateUserService.execute({ email, password })
+
+    return response.json({ user, token })
+  }
+
+}
