+++ backend/src/modules/users/infra/http/controllers/UserAvatarController.ts
@@ -1,0 +1,19 @@
+import { Request, Response } from 'express'
+import { container as dependencyInjector } from 'tsyringe'
+
+import UpdateUserAvatarService from '@/modules/users/services/UpdateUserAvatarService'
+
+export default class UserAvatarController {
+
+  public async update(request: Request, response: Response): Promise<Response> {
+    const updatedUserAvatar = dependencyInjector.resolve(UpdateUserAvatarService)
+
+    const user = await updatedUserAvatar.execute({
+      user_id: request.user.id,
+      avatarFilename: request.file.filename
+    })
+
+    return response.json(user)
+  }
+
+}
