+++ backend/src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,18 +1,9 @@
 import { Router } from 'express'
-import { container as dependencyInjector } from 'tsyringe'
-
-import AuthenticateUserService from '@/modules/users/services/AuthenticateUserService'
+import SessionsController from '../controllers/SessionsController'
 
 const sessionsRouter = Router()
+const sessionsController = new SessionsController()
 
-sessionsRouter.post('/', async (request, response) => {
-    const { email, password } = request.body
-
-    const authenticateUserService = dependencyInjector.resolve(AuthenticateUserService)
-
-    const { user, token } = await authenticateUserService.execute({ email, password })
-
-    return response.json({ user, token })
-})
+sessionsRouter.post('/', sessionsController.create)
 
 export default sessionsRouter
