+++ backend/src/modules/users/infra/http/routes/users.routes.ts
@@ -1,39 +1,25 @@
 import { Router } from 'express'
 import multer from 'multer'
-import { container as dependencyInjector } from 'tsyringe'
 
 import uploadConfig from '@/config/upload'
-import usersTransformer from '@/modules/users/infra/http/transformers/users.transformer'
-import CreateUserService from '@/modules/users/services/CreateUserService'
+
 import ensureAuthenticated from '@/modules/users/infra/http/middlewares/ensureAuthenticated'
-import UpdateUserAvatarService from '@/modules/users/services/UpdateUserAvatarService'
+
+import UsersController from '../controllers/UsersController'
+import UserAvatarController from '../controllers/UserAvatarController'
 
 const usersRouter = Router()
+const usersController = new UsersController()
+const userAvatarController = new UserAvatarController()
 const upload = multer(uploadConfig)
 
-usersRouter.post('/', async (request, response) => {
-    const { name, email, password } = request.body
+usersRouter.post('/', usersController.create)
 
-    const createUserService = dependencyInjector.resolve(CreateUserService)
-
-    const user = await createUserService.execute({ name, email, password })
-
-    return response.json(usersTransformer.renderOne(user))
-})
-
-
-usersRouter.patch('/avatar', ensureAuthenticated, upload.single('avatar'),
-  async (request, response) => {
-
-      const updatedUserAvatar = dependencyInjector.resolve(UpdateUserAvatarService)
-
-      const user = await updatedUserAvatar.execute({
-        user_id: request.user.id,
-        avatarFilename: request.file.filename
-      })
-
-      return response.json(user)
-  }
+usersRouter.patch(
+  '/avatar',
+  ensureAuthenticated,
+  upload.single('avatar'),
+  userAvatarController.update
 )
 
 export default usersRouter
